requirements.txt
# FastAPI y dependencias principales
fastapi>=0.110.0,<1.0.0
starlette>=0.36.3,<1.0.0
uvicorn[standard]>=0.22.0,<1.0.0
python-multipart>=0.0.5

# Pydantic y configuración
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=0.19.0

# Base de datos
sqlalchemy>=2.0.14,<2.1.0
sqlmodel>=0.0.14,<0.1.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.0

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoreo y métricas
prometheus-client>=0.22.0
sentry-sdk>=1.39.0

# Utilidades
structlog>=24.1.0
slowapi>=0.1.8
redis>=5.0.0
